note-x/requirements.txt
Flask
Celery
redis
langchain
openai
markdown
weasyprint
python-dotenv
pdfkit
google-generativeai
gunicorn
